Add a cycle queue variable to rack
The variable will increment whenever a cycle is triggered.
If, at any pointer, during a complete cycle there is another
cycle triggered within the cycle thread, the effect won't
propagate. So if this occurs, another cycle is registered
and then the thread will loop until all the queued cycles
have been performed and go back to waiting on the condition
variable