Add code for elevating the priority of the cycle
The cycle thread in a low latency profile will need to
be woken up very fast. After a lot of staring at the
microsecond latency -- the current solution is to run
the thread in an elevated status which will stop any
very slow wake ups. This code is not cross platform!